feat: read scan directory
