+++ docker-compose.yml
@@ -1,0 +1,10 @@
+version: '3.8'
+name: 'scrutiny-cspm'
+
+services:
+  frontend:
+    build: ./frontend
+    ports:
+      - "3000:3000"
+    volumes:
+      - ./scans:/scans
